module_14/telegram_bot/requirements.txt
aiofiles==24.1.0
aiogram==2.25.1
aiohappyeyeballs==2.4.3
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pyaes==1.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
typing_extensions==4.12.2
yarl==1.17.2
